requirements.txt
pymongo
requests
beautifulsoup4
python-dotenv

